Took another pass on all the test functions. Globals are removed and reorganized into utility classes.
