+++ test/common/bitmap_test.cpp
@@ -25,13 +25,13 @@ TEST(BitmapTests, ByteMultipleCorrectnessTest) {
 
   // Randomly permute bitmap and STL bitmap and compare equality
   std::bitset<num_elements_aligned> stl_bitmap;
-  CheckReferenceBitmap<common::RawBitmap, num_elements_aligned>(*aligned_bitmap, stl_bitmap);
+  ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements_aligned>(*aligned_bitmap, stl_bitmap);
   for (uint32_t i = 0; i < num_iterations; ++i) {
     auto element =
         std::uniform_int_distribution(0, (int) num_elements_aligned - 1)(generator);
     aligned_bitmap->Flip(element);
     stl_bitmap.flip(element);
-    CheckReferenceBitmap<common::RawBitmap, num_elements_aligned>(*aligned_bitmap, stl_bitmap);
+    ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements_aligned>(*aligned_bitmap, stl_bitmap);
   }
 
   common::RawBitmap::Deallocate(aligned_bitmap);
@@ -54,13 +54,13 @@ TEST(BitmapTests, NonByteMultipleCorrectnessTest) {
 
   // Randomly permute bitmap and STL bitmap and compare equality
   std::bitset<num_elements_unaligned> stl_bitmap;
-  CheckReferenceBitmap<common::RawBitmap, num_elements_unaligned>(*non_multiple_bitmap, stl_bitmap);
+  ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements_unaligned>(*non_multiple_bitmap, stl_bitmap);
   for (uint32_t i = 0; i < num_iterations; ++i) {
     auto element =
         std::uniform_int_distribution(0, (int) num_elements_unaligned - 1)(generator);
     non_multiple_bitmap->Flip(element);
     stl_bitmap.flip(element);
-    CheckReferenceBitmap<common::RawBitmap, num_elements_unaligned>(*non_multiple_bitmap, stl_bitmap);
+    ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements_unaligned>(*non_multiple_bitmap, stl_bitmap);
   }
 
   common::RawBitmap::Deallocate(non_multiple_bitmap);
@@ -95,13 +95,13 @@ TEST(BitmapTests, WordUnalignedCorrectnessTest) {
 
   // Randomly permute bitmap and STL bitmap and compare equality
   std::bitset<num_elements> stl_bitmap;
-  CheckReferenceBitmap<common::RawBitmap, num_elements>(*unaligned_bitmap, stl_bitmap);
+  ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements>(*unaligned_bitmap, stl_bitmap);
   for (uint32_t i = 0; i < num_iterations; ++i) {
     auto element =
         std::uniform_int_distribution(0, (int) num_elements - 1)(generator);
     unaligned_bitmap->Flip(element);
     stl_bitmap.flip(element);
-    CheckReferenceBitmap<common::RawBitmap, num_elements>(*unaligned_bitmap, stl_bitmap);
+    ContainerTestUtil::CheckReferenceBitmap<common::RawBitmap, num_elements>(*unaligned_bitmap, stl_bitmap);
   }
 
   delete[] allocated_buffer;
